more windows and sky added
We have added more windows and have found out how to make a rectangle! I didn't think that would be as difficult as it turned out to be but oh well. We're having some issues with adding color options, and have asked in the discord for some assistance. In the meantime, comments have been made in place of future edits.